Initial commit of export action
This is taken from the steps that we have copy-pasted between
Threadbare, Moddable Platformer, and others.

It has been adjusted slightly to support exporting with unstable releases of Godot. To do this,
we must fetch the build from https://downloads.godotengine.org which
redirects to the correct GitHub release; stable releases are on
https://github.com/godotengine/godot but unstable releases are published
to https://github.com/godotengine/godot-builds.